+++ src/Twilio/Rest/Supersim/V1/FleetOptions.php
@@ -22,8 +22,8 @@ abstract class FleetOptions {
      * @param bool $dataEnabled Defines whether SIMs in the Fleet are capable of
      *                          using data connectivity
      * @param int $dataLimit The total data usage (download and upload combined) in
-     *                       Megabytes that each Sim resource assigned to the Fleet
-     *                       resource can consume
+     *                       Megabytes that each Super SIM resource assigned to the
+     *                       Fleet can consume
      * @param string $ipCommandsUrl The URL that will receive a webhook when a
      *                              Super SIM in the Fleet is used to send an IP
      *                              Command from your device
@@ -69,10 +69,13 @@ abstract class FleetOptions {
      * @param string $smsCommandsMethod A string representing the HTTP method to
      *                                  use when making a request to
      *                                  `sms_commands_url`
+     * @param int $dataLimit The total data usage (download and upload combined) in
+     *                       Megabytes that each Super SIM assigned to the Fleet
+     *                       can consume
      * @return UpdateFleetOptions Options builder
      */
-    public static function update(string $uniqueName = Values::NONE, string $networkAccessProfile = Values::NONE, string $ipCommandsUrl = Values::NONE, string $ipCommandsMethod = Values::NONE, string $smsCommandsUrl = Values::NONE, string $smsCommandsMethod = Values::NONE): UpdateFleetOptions {
-        return new UpdateFleetOptions($uniqueName, $networkAccessProfile, $ipCommandsUrl, $ipCommandsMethod, $smsCommandsUrl, $smsCommandsMethod);
+    public static function update(string $uniqueName = Values::NONE, string $networkAccessProfile = Values::NONE, string $ipCommandsUrl = Values::NONE, string $ipCommandsMethod = Values::NONE, string $smsCommandsUrl = Values::NONE, string $smsCommandsMethod = Values::NONE, int $dataLimit = Values::NONE): UpdateFleetOptions {
+        return new UpdateFleetOptions($uniqueName, $networkAccessProfile, $ipCommandsUrl, $ipCommandsMethod, $smsCommandsUrl, $smsCommandsMethod, $dataLimit);
     }
 }
 
@@ -83,8 +86,8 @@ class CreateFleetOptions extends Options {
      * @param bool $dataEnabled Defines whether SIMs in the Fleet are capable of
      *                          using data connectivity
      * @param int $dataLimit The total data usage (download and upload combined) in
-     *                       Megabytes that each Sim resource assigned to the Fleet
-     *                       resource can consume
+     *                       Megabytes that each Super SIM resource assigned to the
+     *                       Fleet can consume
      * @param string $ipCommandsUrl The URL that will receive a webhook when a
      *                              Super SIM in the Fleet is used to send an IP
      *                              Command from your device
@@ -136,11 +139,11 @@ class CreateFleetOptions extends Options {
     }
 
     /**
-     * The total data usage (download and upload combined) in Megabytes that each Sim resource assigned to the Fleet resource can consume during a billing period (normally one month). Value must be between 1MB (1) and 2TB (2,000,000). Defaults to 1GB (1,000).
+     * The total data usage (download and upload combined) in Megabytes that each Super SIM assigned to the Fleet can consume during a billing period (normally one month). Value must be between 1MB (1) and 2TB (2,000,000). Defaults to 1GB (1,000).
      *
      * @param int $dataLimit The total data usage (download and upload combined) in
-     *                       Megabytes that each Sim resource assigned to the Fleet
-     *                       resource can consume
+     *                       Megabytes that each Super SIM resource assigned to the
+     *                       Fleet can consume
      * @return $this Fluent Builder
      */
     public function setDataLimit(int $dataLimit): self {
@@ -272,14 +275,18 @@ class UpdateFleetOptions extends Options {
      * @param string $smsCommandsMethod A string representing the HTTP method to
      *                                  use when making a request to
      *                                  `sms_commands_url`
+     * @param int $dataLimit The total data usage (download and upload combined) in
+     *                       Megabytes that each Super SIM assigned to the Fleet
+     *                       can consume
      */
-    public function __construct(string $uniqueName = Values::NONE, string $networkAccessProfile = Values::NONE, string $ipCommandsUrl = Values::NONE, string $ipCommandsMethod = Values::NONE, string $smsCommandsUrl = Values::NONE, string $smsCommandsMethod = Values::NONE) {
+    public function __construct(string $uniqueName = Values::NONE, string $networkAccessProfile = Values::NONE, string $ipCommandsUrl = Values::NONE, string $ipCommandsMethod = Values::NONE, string $smsCommandsUrl = Values::NONE, string $smsCommandsMethod = Values::NONE, int $dataLimit = Values::NONE) {
         $this->options['uniqueName'] = $uniqueName;
         $this->options['networkAccessProfile'] = $networkAccessProfile;
         $this->options['ipCommandsUrl'] = $ipCommandsUrl;
         $this->options['ipCommandsMethod'] = $ipCommandsMethod;
         $this->options['smsCommandsUrl'] = $smsCommandsUrl;
         $this->options['smsCommandsMethod'] = $smsCommandsMethod;
+        $this->options['dataLimit'] = $dataLimit;
     }
 
     /**
@@ -354,6 +361,19 @@ class UpdateFleetOptions extends Options {
      */
     public function setSmsCommandsMethod(string $smsCommandsMethod): self {
         $this->options['smsCommandsMethod'] = $smsCommandsMethod;
+        return $this;
+    }
+
+    /**
+     * The total data usage (download and upload combined) in Megabytes that each Super SIM assigned to the Fleet can consume during a billing period (normally one month). Value must be between 1MB (1) and 2TB (2,000,000). Defaults to 1GB (1,000).
+     *
+     * @param int $dataLimit The total data usage (download and upload combined) in
+     *                       Megabytes that each Super SIM assigned to the Fleet
+     *                       can consume
+     * @return $this Fluent Builder
+     */
+    public function setDataLimit(int $dataLimit): self {
+        $this->options['dataLimit'] = $dataLimit;
         return $this;
     }