+++ tests/Twilio/Integration/Supersim/V1/FleetTest.php
@@ -42,7 +42,7 @@ class FleetTest extends HolodeckTestCase {
                 "account_sid": "ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "unique_name": "unique_name",
                 "data_enabled": true,
-                "data_limit": 1000,
+                "data_limit": 500,
                 "data_metering": "payg",
                 "date_created": "2019-07-30T20:00:00Z",
                 "date_updated": "2019-07-30T20:00:00Z",
@@ -200,15 +200,15 @@ class FleetTest extends HolodeckTestCase {
         ));
     }
 
-    public function testUpdateUniqueNameResponse(): void {
+    public function testUpdateUniqueNameAndDataLimitResponse(): void {
         $this->holodeck->mock(new Response(
             200,
             '
             {
                 "account_sid": "ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
-                "unique_name": "unique_name",
+                "unique_name": "GPS Trackers Fleet 5GB",
                 "data_enabled": true,
-                "data_limit": 1000,
+                "data_limit": 5000,
                 "data_metering": "payg",
                 "date_created": "2019-10-15T20:00:00Z",
                 "date_updated": "2019-10-15T20:00:00Z",